bot/requirements.txt
discord.py
python-dotenv
PyNaCl
aiohttp
loguru